// frontend/src/enums/section.enu.ts
export enum ESection {
  PROFILE = "profile",
  WORK = "workHistory",
  EDUCATION = "education",
  SKILLS = "skills",
  SUMMARY = "summary",
  HOBBIES = "hobbies",
  CERTIFICATES = "certification",
  LANGUAGES = "languages",
  ACHIEVEMENTS = "achievements",
  REFERENCES = "references",
  PUBLICATIONS = "publications",
  CUSTOMSECTION = "customSection",
}